added check of file

+++ BuildBotLib/asssetsinstaller.py
@@ -18,7 +18,8 @@ def NDKDownloadCMD(props):
     format = link[link.rfind('.'):]
 
     if module == "AndroidNDK":
-        os.remove("temp" + format)
+        if os.path.isfile(module):
+            os.remove("temp" + format)
         res = ["curl", link, "--output", "temp" + format]
 
     return res
